// frontend/src/app/model/produto.ts
export class Produto {
    codigo: number = 0;
    nome: string = "";
    descritivo: string = "";
    valor: number = 0;
    quantidade: number = 0;
    valorPromo: number = 0;
    categoria: string = ""; // Ex: escolar, escritório, arte
    imagem: string = ""; // URL da imagem
    destaque: number = 0;
}